deleted old server files and added default port for client server
