+++ client/vite.config.js
@@ -7,5 +7,8 @@ export default defineConfig(() => {
             outDir: 'build',
         },
         plugins: [react()],
+        server: {
+            port: 3000,
+        },
     };
 });
